views to components folder
