+++ 3o14/src/components/actorlink.tsx
@@ -1,0 +1,26 @@
+import type { FC } from "hono/jsx";
+import type { Actor } from "../models/actor.ts";
+
+export interface ActorLinkProps {
+  actor: Actor;
+}
+
+export const ActorLink: FC<ActorLinkProps> = ({ actor }) => {
+  const href = actor.url ?? actor.uri;
+  return actor.name == null ? (
+    <a href={href} class="secondary">
+      {actor.handle}
+    </a>
+  ) : (
+    <>
+      <a href={href}>{actor.name}</a>{" "}
+      <small>
+        (
+        <a href={href} class="secondary">
+          {actor.handle}
+        </a>
+        )
+      </small>
+    </>
+  );
+};
