+++ 3o14/src/components/index.ts
@@ -1,0 +1,10 @@
+export * from "./actorlink.tsx";
+export * from "./followerlist.tsx";
+export * from "./followinglist.tsx";
+export * from "./home.tsx";
+export * from "./layout.tsx";
+export * from "./postlist.tsx";
+export * from "./postpage.tsx";
+export * from "./postview.tsx";
+export * from "./profile.tsx";
+export * from "./setupform.tsx";
