+++ 3o14/src/components/layout.tsx
@@ -1,0 +1,19 @@
+import type { FC } from "hono/jsx";
+
+export const Layout: FC = (props) => (
+  <html lang= "en" >
+  <head>
+  <meta charset="utf-8" />
+    <meta name="viewport" content = "width=device-width, initial-scale=1" />
+      <meta name="color-scheme" content = "light dark" />
+        <title>3o14 </title>
+          < link
+rel = "stylesheet"
+href = "https://cdn.jsdelivr.net/npm/@picocss/pico@2/css/pico.min.css"
+  />
+  </head>
+  < body >
+  <main class="container" > { props.children } </main>
+    </body>
+    </html>
+);
