+++ 3o14/src/components/setupform.tsx
@@ -1,0 +1,25 @@
+import type { FC } from "hono/jsx";
+
+export const SetupForm: FC = () => (
+  <>
+    <h1>Set up your microblog</h1>
+    <form method="post" action="/setup">
+      <fieldset>
+        <label>
+          Username{" "}
+          <input
+            type="text"
+            name="username"
+            required
+            maxlength={50}
+            pattern="^[a-z0-9_\-]+$"
+          />
+        </label>
+        <label>
+          Name <input type="text" name="name" required />
+        </label>
+      </fieldset>
+      <input type="submit" value="Setup" />
+    </form>
+  </>
+);
